Extract classes and tests into different files. Put game_over function in GameBoard class. Replace assert not by refute
