+++ test/test_helper.rb
@@ -3,6 +3,8 @@ require 'bundler/setup'
 require 'minitest/autorun'
 
 require 'connect_four'
+require 'game'
+require 'game_board'
 
 class ConnectFourSpec < Minitest::Test
 end
